change from environment-name to environment-key (#14)
# Summary
Use environment-key instead of environment-name in the environment mapping command

https://www.notion.so/statsig/LD-Migration-Script-Improvements-246d27b5090d80b593cde1f16ebb17ea?source=copy_link#258d27b5090d80a395eff57f144d1c25

# Test

See that the script accepts the `environment-key-mapping` param.

![image.png](https://app.graphite.dev/user-attachments/assets/2a67a55d-730d-4a9a-9917-977da5fabc89.png)

Map LD test environment to Statsig staging. On console, see that all configs on LD test environment are mapped to Statsig staging (i.e the rules that are prefixed with `test` are active on staging)

![image.png](https://app.graphite.dev/user-attachments/assets/6263e13d-7a29-4b3f-9950-314800fc0bce.png)