delete configs based on dependencies (#23)
# summary

delete configs in statsig based on dependencies: all dcs in any order -> gates by order of dependencies -> segments by order of dependencies

e.g if gateA has a rule that refers to gateB, we want to delete gateA before gateB. same for segments

# test

1. have a dc depends on gate 1 depends on gate 2 depends on segment on statsig. have a segment depends on another segment.
2. put the id of these configs in list of `validConfigNames` in `cli`. 
3. put the tag "Imported from launchdarkly" on them.

run the script. expect all the configs to be deleted.