refactor importConfigs to return a list of imported result instead of summaries (#31)
# Summary

`importConfigs` is returning a summary of the number of imported configs, along with notices and errors. Instead, we need the function to return every import result for every config, so that in the next PR, we can log each import result to a CSV file.

# Test
Run the script and see that the terminal remains the same.

![image.png](https://app.graphite.dev/user-attachments/assets/d33373b0-82ed-468d-a92f-d4f4ffa6b2e8.png)