create custom field based on heuristic if no custom attribute mapping (#32)
# Summary

If there's a custom attribute in LD that is not explicitly mapped to a custom field (i.e user didn't pass in a context attribute to custom field mapping for that custom attribute), we use heuristic to map it. 

E.g `organization` context kind, `location` custom attribute will be mapped to `organization.location`

# Test

I commented out the mapping between LD `user.ip` and Statsig `ip_address` so the script can treat `user.ip` as an unmapped custom attribute.

Have a rule in LD that uses `user.ip`

![image.png](https://app.graphite.dev/user-attachments/assets/d3f04a3d-84a1-4ef9-a01c-5d4cad8bdd3b.png)

Run the script.

Expect a notice in the terminal.

![image.png](https://app.graphite.dev/user-attachments/assets/0cf24032-097b-4d1a-ae13-423feb7bbb7e.png)

Expect custom field `user.ip` in Statsig

![image.png](https://app.graphite.dev/user-attachments/assets/1d29e200-ddd6-4925-9549-71d0e49a6b8d.png)