log company using script and errors (#35)
# Summary

- Set up logging function
- Log the company using the script
- Log the errors the user encounter with the script

# Test

Run the script without passing all the required arguments. Check the Log Stream of the Statsig project. Expect to see `migration_script_error` events with value as the error message and metadata includes all the info about the company running the script.

Run the script again. When asked a yes/no question, answer anything but `y`. Check the Log Stream. Expect to see `migration_script_interrupt`.

To make the demo clearer, I locally swapped the client key with that of one of my projects:

[Screen Recording 2025-09-18 at 5.23.22 PM.mov <span class="graphite__hidden">(uploaded via Graphite)</span> <img class="graphite__hidden" src="https://app.graphite.dev/user-attachments/thumbnails/8c235c01-98aa-411c-a6db-a2c9d4e9f411.mov" />](https://app.graphite.dev/user-attachments/video/8c235c01-98aa-411c-a6db-a2c9d4e9f411.mov)