log final result of the migration (#37)
# Summary

Logging when the script succeeds

# Test

Run the script and ensure that it succeeds. Go to Event Log Stream, expect the `migration_script_success` event with information about the number of flags imported/not imported and all the errors and notices generated throughout the script.

To make the demo clearer, I locally swapped the client key with that of one of my projects:

[Screen Recording 2025-09-19 at 9.53.17 AM.mov <span class="graphite__hidden">(uploaded via Graphite)</span> <img class="graphite__hidden" src="https://app.graphite.dev/user-attachments/thumbnails/fd7d1007-647c-49b6-bca3-b81ab9f7c771.mov" />](https://app.graphite.dev/user-attachments/video/fd7d1007-647c-49b6-bca3-b81ab9f7c771.mov)