map the context kind user to whatever the user specifies instead of user_id (#43)
# Summary

We automatically map context kind `user` to `user_id` and there's nothing the user can do about it... until now. Now we map the context kind `user` to whatever the user specifies. The fallback is still `user_id`.

[Linear](https://linear.app/statsig/issue/FM-1004/make-user-context-kind-mappable-instead-of-ignoring-it)

# Test

Have a flag in LD that uses context kind `user` for both rule condition and randomization.

![image.png](https://app.graphite.dev/user-attachments/assets/4e26be51-52f9-430d-acb9-5511fd94a4e8.png)

In the script, map the context kind `user` to `organizationID`. Make sure the statsig project has the organizationID first. Run the script. Expect to see the context kind is converted to organizationID in Statsig.

![image.png](https://app.graphite.dev/user-attachments/assets/60c6af36-bf0b-4821-92ef-2a787e2f386b.png)

Run the script again without explicitly mapping context kind `user` to anything. Expect it to be mapped to user id in statsig.

![image.png](https://app.graphite.dev/user-attachments/assets/ba5e53cc-c76c-4a8b-b666-5a579b2f10e2.png)